# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CampusSummary {
  buildings: number;
  rooms: number;
  activeEvents: number;
  liveZones: number;
}

export type BuildingPosition = {
  x: number;
  y: number;
};

export interface Building {
  id: string;
  code: string;
  name: string;
  floors: number;
  accent: string;
  position: BuildingPosition;
}

export interface CampusPlace {
  id: string;
  name: string;
  type: string;
  buildingId: string;
  floor: number;
  nodeId: string;
  accessible: boolean;
  x: number;
  y: number;
  /** @nullable */
  capacity?: number | null;
  /** @nullable */
  details?: string | null;
}

export type PlaceStatusStatus = typeof PlaceStatusStatus[keyof typeof PlaceStatusStatus];


export const PlaceStatusStatus = {
  available: 'available',
  moderate: 'moderate',
  busy: 'busy',
  closed: 'closed',
} as const;

export interface PlaceStatus {
  placeId: string;
  current: number;
  capacity: number;
  status: PlaceStatusStatus;
  updatedAt: string;
}

export interface RouteInput {
  startNodeId: string;
  destinationNodeId: string;
  accessibleOnly?: boolean;
}

export type RouteResponseStepsItem = {
  label: string;
  detail: string;
  kind: string;
};

export interface RouteResponse {
  distanceMeters: number;
  durationMinutes: number;
  floors: number[];
  accessible: boolean;
  steps: RouteResponseStepsItem[];
}

export type SearchPlacesParams = {
q?: string;
buildingId?: string;
floor?: number;
};