# lib/api-zod/src/generated/types/building.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BuildingPosition } from './buildingPosition';

export interface Building {
  id: string;
  code: string;
  name: string;
  floors: number;
  accent: string;
  position: BuildingPosition;
}