# lib/api-zod/src/generated/types/campusPlace.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CampusPlace {
  id: string;
  name: string;
  type: string;
  buildingId: string;
  floor: number;
  nodeId: string;
  accessible: boolean;
  x: number;
  y: number;
  /** @nullable */
  capacity?: number | null;
  /** @nullable */
  details?: string | null;
}