# lib/api-zod/src/generated/types/campusSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CampusSummary {
  buildings: number;
  rooms: number;
  activeEvents: number;
  liveZones: number;
}