# lib/api-zod/src/generated/types/placeStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PlaceStatusStatus } from './placeStatusStatus';

export interface PlaceStatus {
  placeId: string;
  current: number;
  capacity: number;
  status: PlaceStatusStatus;
  updatedAt: string;
}