# lib/api-zod/src/generated/types/routeInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RouteInput {
  startNodeId: string;
  destinationNodeId: string;
  accessibleOnly?: boolean;
}