# lib/api-zod/src/generated/types/routeResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RouteResponseStepsItem } from './routeResponseStepsItem';

export interface RouteResponse {
  distanceMeters: number;
  durationMinutes: number;
  floors: number[];
  accessible: boolean;
  steps: RouteResponseStepsItem[];
}